package name added, mantid logger replaced with python logger, environment updated
