+++ src/hyspecplanningtools/configuration.py
@@ -8,12 +8,12 @@ import shutil
 from configparser import ConfigParser
 from pathlib import Path
 
-from mantid.kernel import Logger
+import logging
 
-logger = Logger("PACKAGENAME")
+logger = logging.getLogger("hyspecplanningtools")
 
 # configuration settings file path
-CONFIG_PATH_FILE = os.path.join(Path.home(), ".packagename", "configuration.ini")
+CONFIG_PATH_FILE = os.path.join(Path.home(), ".hyspecplanningtools", "configuration.ini")
 
 
 class Configuration:
@@ -30,7 +30,7 @@ class Configuration:
 
         # retrieve the file path of the file
         self.config_file_path = CONFIG_PATH_FILE
-        logger.information(f"{self.config_file_path} will be used")
+        logger.info(f"{self.config_file_path} will be used")
 
         # if template conf file path exists
         if os.path.exists(self.template_file_path):
