+++ src/hyspecplanningtools/configuration_template.ini
